requirements.txt
anthropic
apscheduler
requests
python-dotenv
